tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/WindowTimers/003.html

<!--
setInterval(function() { postMessage(1); }, 10);
/*
-->
<!doctype html>
<title>setInterval</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, 1);
    this.done();
  });
});
</script>
<!--
*/
//-->
